Separating the main gl_program functionality
I added a base class for all GL components.
The repeated functionality is now in this base class. In the painter code
normally only the shader code have to be given.
